Returned donations is supported * Did some refactoring * Moved code to separate files
